requirements.txt
bleach>=3.1.1,<5.0
celery[redis]>=5.0.3,<5.3
charset-normalizer>=2.0,<3.0
cssselect>=1.0.0,<1.2
Cython>=0.29.14,<0.30
diff-match-patch==20200713
django-appconf>=1.0.3,<1.1
django-compressor>=2.4,<3.2
django-crispy-forms>=1.9.0,<1.15
django-filter>=2.4.0,<22.0
django-redis>=4.11.0,<6.0
Django[argon2]>=3.2,<4.1
djangorestframework>=3.11.0,<3.14
filelock>=3.2.0,<3.7
GitPython>=2.1.15,<3.2
hiredis>=1.0.1,<2.1
html2text>=2019.8.11,<2020.1.17
importlib-metadata>=3.6.0,<4.12; python_version < '3.8'
jellyfish>=0.7.2,<0.10,!=0.8.3,!=0.8.4,!=0.8.5,!=0.8.6,!=0.8.7,!=0.8.8
jsonschema>=3.0.0,<4.5
lxml>=4.6.3,<4.9
openpyxl>=2.6.0,<3.1,!=3.0.2
Pillow>=6.0.0,<10.0
pyahocorasick>=1.4,<1.5
Pygments>=2.5.1,<3.0
pyicumessageformat>=1.0.0,<1.1
pyparsing>=3.0.5,<3.1
python-dateutil>=2.8.1
python-redis-lock>=3.6.0,<3.8
requests>=2.26.0,<2.28
sentry_sdk>=0.13.0,<1.6
siphashc>=1.2,<3.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
translate-toolkit>=3.6.0,<3.8
translation-finder>=2.7,<3.0
user-agents>=2.0,<2.3
weblate-language-data>=2021.5
weblate-schemas==0.7
